pyproject.toml
[project]
name = "pashnev-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "abstractservice",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.5",
    "fastapi[standard]>=0.135.1",
    "motor>=3.7.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
]
